+++ common/src/main/java/club/iananderson/pocketgps/PocketGps.java
@@ -1,7 +1,11 @@
 package club.iananderson.pocketgps;
 
+import club.iananderson.pocketgps.config.PocketGpsConfig;
+import club.iananderson.pocketgps.energy.EnergyUnit;
 import club.iananderson.pocketgps.platform.Services;
+import java.util.function.Supplier;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.Item;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -9,9 +13,20 @@ public final class PocketGps {
   public static final String MOD_ID = "pocketgps";
   public static final String MOD_NAME = "PocketGPS";
   public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
+  public static String ENERGY_TAG = "energy";
+  public static String TOGGLE_GPS_TAG = "toggle_gps";
+  public static Supplier<Item> GPS;
+  public static ResourceLocation TOGGLE_GPS = new ResourceLocation(MOD_ID, TOGGLE_GPS_TAG);
   private static boolean accessoriesLoaded;
   private static boolean curiosLoaded;
-  private static String platformName;
+  private static boolean trinketsLoaded;
+  private static boolean worldMapLoaded;
+  private static EnergyUnit energyUnit;
+  private static boolean gpsNeedPower;
+  private static int gpsEnergyCapacity;
+  private static int gpsMaxInput;
+  private static int gpsMaxOutput;
+  private static int gpsEnergyCost;
 
   private PocketGps() {
   }
@@ -22,8 +37,19 @@ public final class PocketGps {
 
   public static void init() {
     platformName = Services.PLATFORM.getPlatformName();
-    curiosLoaded = Services.PLATFORM.isModLoaded("trinkets") || Services.PLATFORM.isModLoaded("curios");
+    curiosLoaded = Services.PLATFORM.isModLoaded("curios");
+    trinketsLoaded = Services.PLATFORM.isModLoaded("trinkets");
     accessoriesLoaded = Services.PLATFORM.isModLoaded("accessories");
+    worldMapLoaded = Services.PLATFORM.isModLoaded("xaeroworldmap");
+  }
+
+  public static void clientInit() {
+    energyUnit = PocketGpsConfig.getEnergyUnit();
+    gpsNeedPower = PocketGpsConfig.getGpsNeedPower();
+    gpsEnergyCapacity = PocketGpsConfig.getGpsEnergyCapacity();
+    gpsMaxInput = PocketGpsConfig.getGpsMaxInput();
+    gpsMaxOutput = PocketGpsConfig.getGpsMaxOutput();
+    gpsEnergyCost = PocketGpsConfig.getGpsEnergyCost();
   }
 
   public static String platformName() {
@@ -31,10 +57,46 @@ public final class PocketGps {
   }
 
   public static boolean curiosLoaded() {
-    return PocketGps.curiosLoaded;
+    return curiosLoaded;
+  }
+
+  public static boolean trinketsLoaded() {
+    return trinketsLoaded;
   }
 
   public static boolean accessoriesLoaded() {
-    return PocketGps.accessoriesLoaded;
+    return accessoriesLoaded;
+  }
+
+  public static boolean worldMapLoaded() {
+    return worldMapLoaded;
+  }
+
+  public static EnergyUnit energyUnit() {
+    return energyUnit;
+  }
+
+  public static boolean gpsNeedPower() {
+    return gpsNeedPower;
+  }
+
+  public static int gpsEnergyCapacity() {
+    return gpsEnergyCapacity;
+  }
+
+  public static int gpsMaxInput() {
+    return gpsMaxInput;
+  }
+
+  public static int gpsMaxOutput() {
+    return gpsMaxOutput;
+  }
+
+  public static int gpsEnergyCost() {
+    return gpsEnergyCost;
+  }
+
+  public static ResourceLocation location(String path) {
+    return new ResourceLocation(MOD_ID, path);
   }
 }