+++ fabric/src/main/java/club/iananderson/pocketgps/fabric/PocketGpsFabric.java
@@ -1,18 +1,23 @@
 package club.iananderson.pocketgps.fabric;
 
 import club.iananderson.pocketgps.PocketGps;
-import club.iananderson.pocketgps.accessories.AccessoriesCompat;
+import club.iananderson.pocketgps.client.PocketGpsClient;
 import club.iananderson.pocketgps.fabric.registry.FabricRegistration;
+import club.iananderson.pocketgps.impl.accessories.AccessoriesCompat;
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.entity.event.v1.ServerPlayerEvents;
 
 public final class PocketGpsFabric implements ModInitializer {
+  //Todo ---- Find out why it wont receive power on servers only
+
   @Override
   public void onInitialize() {
     PocketGps.init();
-    FabricRegistration.init();
-    if (PocketGps.accessoriesLoaded()) {
+    FabricRegistration.register();
+
+    if (PocketGps.accessoriesLoaded() && !PocketGps.trinketsLoaded()) {
       PocketGps.LOG.info("Talking to Accessories");
-      AccessoriesCompat.init(FabricRegistration.POCKET_GPS);
+      AccessoriesCompat.init(PocketGps.GPS.get());
     }
   }
 }