+++ fabric/src/main/java/club/iananderson/pocketgps/fabric/registry/FabricRegistration.java
@@ -1,19 +1,43 @@
 package club.iananderson.pocketgps.fabric.registry;
 
 import club.iananderson.pocketgps.PocketGps;
-import club.iananderson.pocketgps.items.PocketGpsItems;
+import club.iananderson.pocketgps.fabric.items.ChargeableGpsItem;
+import club.iananderson.pocketgps.registry.CommonRegistration;
+import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
+import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.CreativeModeTab;
-import net.minecraft.world.item.Item;
 
 public class FabricRegistration {
+
   public static final Item POCKET_GPS = PocketGpsItems.GPS_ITEM_SUPPLIER.get();
+  private static final ResourceKey<CreativeModeTab> ITEM_GROUP;
 
-  public static CreativeModeTab TAB = PocketGpsItems.POCKET_GPS_TAB(POCKET_GPS).get();
+  public static ChargeableGpsItem POCKET_GPS = new ChargeableGpsItem();
 
-  public static void init() {
-    Registry.register(BuiltInRegistries.ITEM, PocketGps.location("gps"), POCKET_GPS);
+  public static CreativeModeTab TAB = TABS.title(Component.translatable("tab.pocketgps"))
+      .icon(() -> CommonRegistration.addIcon(PocketGps.GPS.get())).displayItems((par, out) -> entries(out))
+      .build();
+
+  static {
+    PocketGps.GPS = () -> POCKET_GPS;
+    ITEM_GROUP = ResourceKey.create(Registries.CREATIVE_MODE_TAB, PocketGps.location("item_group"));
+  }
+
+  private static void entries(CreativeModeTab.Output entries) {
+    entries.acceptAll(CommonRegistration.addPoweredItem(PocketGps.GPS.get(), true));
+  }
+
+  public static void register() {
     Registry.register(BuiltInRegistries.CREATIVE_MODE_TAB, PocketGps.location("tab"), TAB);
+    Registry.register(BuiltInRegistries.ITEM, PocketGps.location("gps"), POCKET_GPS);
+
+    ItemGroupEvents.modifyEntriesEvent(ITEM_GROUP)
+        .register(FabricRegistration::entries);
   }
 }