+++ CHANGELOG.md
@@ -4,25 +4,82 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
+## [1.1.5] - 2025-01-24
+
+### Fixed
+
+- Fixed minimaps displaying on initial load into world.
+- Fixed minimaps reappearing on servers when the player dies.
+
+## [1.1.4] - 2025-01-15
+
+### Changed
+
+- Improved logic of detecting power level
+- Made the GPS screen turn off if there is no power
+- Change the energy bar to not show if the GPS is fully charged
+- Reworked the item used for the icon
+
+### Fixed
+
+- If the GPS is set to not use power in the config, the uncharged GPS will no longer be added to the creative tab or 
+  REI/JEI
+
+## [1.1.3] - 2025-01-15
+
+### Fixed
+
+- Fixed crash when Xaero's Worldmap was not installed
+- Fixed typo in config
+
+## [1.1.2] - 2025-01-13
+
+### Fixed
+
+- Fixed crash with Trinkets installed.
+
+## [1.1.1] - 2025-01-12
+
+### Changed
+
+- Updated Russian translation (provided by @mpustovoi)
+
+## [1.1.0] - 2025-01-12
+
+### Added
+
+- Added a config file
+- Added optional energy requirement for the GPS
+- Reworked GPS detection
+- Added support for Xaero's Worldmap
+
+### Changed
+
+- Switched to the built-in effects from Xaero's Minimap/Worldmap for disabling their mini/world maps.
+
 ## [1.0.6] - 2024-11-11
 
 ### Added
+
 - Added Russian translation (provided by @mpustovoi)
 - Built against latest versions of Xaero's Minimap, Journeymap, and FTB Chunks
 
 ## [1.0.5] - 2024-09-27
 
 ### Added
+
 - Added Simplified Chinese translation (via @suoyukii)
 
 ## [1.0.4] - 2024-09-25
 
 ### Fixed
+
 - Fixed crash when dying with GPS equipped in Curio/Trinket/Accessories slot.
 
 ## [1.0.3] - 2024-09-22
 
 ### Fixed
+
 - Fixed crash if run on server
 
 ## [1.0.2] - 2024-09-13